adapter - remove ChannelValidator & clean up code TODO: uncomment sweeper test case once we figure out what's wrong with the contract call
